mathcing engine, renamed folder and cleaned up OrderComparator
